+++ packages/terra-table/CHANGELOG.md
@@ -7,6 +7,7 @@
   
 * Changed
   * Updated the table component so that the cell dive-in logic would not execute when not in the grid context.
+  * Modified the table component so that it can receive focus when scrollable.
 
 ## 5.0.0-alpha.0 - (October 17, 2023)